market_data/requirements.txt
# Market Data Module Dependencies

# Massive (formerly Polygon.io) - Market data API
massive>=1.12.0

# Database
asyncpg>=0.29.0           # PostgreSQL async driver
psycopg2-binary>=2.9.0    # PostgreSQL sync driver (for setup scripts)

# Data validation and models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data processing (optional, for future use)
# pandas>=2.0.0
# numpy>=1.24.0
